WIP proxy request to my own server to inspect path
In prod the proxy is giving me path errors, but they look to be coming
back from Notion. So the proxy is sort of working. The traffic is
getting where it needs to go, but somehow the path is broken. Not sure
what the difference between prod and dev would be in this case, hoping
to find out.